Log protocol instaed of manual vs auto send
