+++ collect_app/src/main/java/org/odk/collect/android/instancemanagement/InstancesDataService.kt
@@ -229,7 +229,6 @@ class InstancesDataService(
             instances,
             referrer,
             overrideURL,
-            false,
             cancelAfterAuthException,
             externalDeleteAfterUpload,
             defaultSuccessMessage,
