+++ collect_app/src/main/java/org/odk/collect/android/instancemanagement/send/InstanceSubmitter.kt
@@ -1,9 +1,5 @@
 package org.odk.collect.android.instancemanagement.send
 
-import org.odk.collect.analytics.Analytics
-import org.odk.collect.android.analytics.AnalyticsEvents
-import org.odk.collect.android.analytics.AnalyticsUtils
-import org.odk.collect.android.application.Collect
 import org.odk.collect.android.instancemanagement.InstanceDeleter
 import org.odk.collect.android.projects.ProjectDependencyModule
 import org.odk.collect.android.utilities.InstanceAutoDeleteChecker
@@ -28,7 +24,6 @@ class InstanceSubmitter(
         toUpload: List<Instance>,
         referrer: String = "",
         overrideURL: String? = null,
-        autoSend: Boolean = true,
         cancelAfterAuthException: Boolean = false,
         externalDeleteAfterUpload: Boolean? = null,
         defaultSuccessMessage: String? = null,
@@ -49,12 +44,10 @@ class InstanceSubmitter(
             onProgress( index + 1, sortedInstances.size)
 
             try {
-                val resultMessage = instanceUploader.uploadOneSubmission(projectId, instance, deviceId, overrideURL)
+                val resultMessage = instanceUploader.uploadOneSubmission(projectId, instance, deviceId, overrideURL, referrer)
                 uploadResults.add(InstanceUploadResult.Success(instance, resultMessage ?: defaultSuccessMessage))
 
                 deleteInstance(instance, formsRepository, instancesRepository, generalSettings, externalDeleteAfterUpload)
-                logOverrideURL(referrer, overrideURL)
-                logUploadedForm(instance, autoSend)
             } catch (e: FormUploadException) {
                 Timber.d(e)
                 uploadResults.add(InstanceUploadResult.Error(instance, e))
@@ -88,23 +81,5 @@ class InstanceSubmitter(
                 formsRepository
             ).delete(instance.dbId)
         }
-    }
-
-    private fun logOverrideURL(referrer: String, overrideURL: String?) {
-        if (overrideURL != null) {
-            Analytics.log(
-                AnalyticsEvents.INSTANCE_UPLOAD_CUSTOM_SERVER,
-                "label",
-                referrer
-            )
-        }
-    }
-
-    private fun logUploadedForm(instance: Instance, autoSend: Boolean) {
-        AnalyticsUtils.logFormEvent(
-            AnalyticsEvents.SUBMISSION,
-            instance.formId,
-            if (autoSend) "HTTP auto" else "HTTP"
-        )
     }
 }
