+++ collect_app/src/main/java/org/odk/collect/android/instancemanagement/send/InstanceUploader.kt
@@ -4,5 +4,5 @@ import org.odk.collect.forms.instances.Instance
 
 interface InstanceUploader {
     @Throws(FormUploadException::class)
-    fun uploadOneSubmission(projectId: String, instance: Instance, deviceId: String?, overrideURL: String?): String?
+    fun uploadOneSubmission(projectId: String, instance: Instance, deviceId: String?, overrideURL: String?, referrer: String): String?
 }
